+++ hcloud/resource_hcloud_floatingip_test.go
@@ -62,7 +62,7 @@ resource "hcloud_server" "floating_ip1" {
   name        = "floating-ip-1-%d"
   server_type = "cx11"
 	image       = "debian-9"
-  datacenter  = "fsn1-dc8"
+  datacenter  = "fsn1-dc14"
   ssh_keys    = ["${hcloud_ssh_key.floating_ip.id}"]
 }
 
@@ -83,7 +83,7 @@ resource "hcloud_server" "floating_ip1" {
   name        = "floating-ip-1-%d"
   server_type = "cx11"
 	image       = "debian-9"
-  datacenter  = "fsn1-dc8"
+  datacenter  = "fsn1-dc14"
   ssh_keys    = ["${hcloud_ssh_key.floating_ip.id}"]
 }