+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## 1.4.0 (Unreleased)
 
+FEATURES:
+
+* **New Resource**: `hcloud_volume` Manage volumes.
+* **New Datasource**: `hcloud_volume` Lookup volumes.
+
 NOTES:
+
 * **Deprecation**: resource/hcloud_server: `backup_window`, will be removed in the near future.
 
 ## 1.3.0 (September 12, 2018)